+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Change log entries should be formulated in the imperative present tense.
 
 ## Unreleased
 
+### Changed
+
+* Add optional sender `IActorRef` parameter to `WhenTelling` methods of `MarbleTestExtensionsForAkka`.
 
 ## [1.0.0] - 2018-12-14